Remove internal namespaces (breaks API-Extractor)
