+++ packages/dds/tree/src/class-tree/index.ts
@@ -40,10 +40,6 @@ export {
 	typedObjectValues,
 } from "./schemaCreationUtilities";
 
-// Below here are things that are used by the above, but not part of the desired API surface.
-import * as InternalClassTreeTypes from "./internal";
-export { InternalClassTreeTypes };
-
 // Exporting the schema (RecursiveObject) to test that recursive types are working correctly.
 // These are `@internal` so they can't be included in the `InternalClassTreeTypes` due to https://github.com/microsoft/rushstack/issues/3639
 export {
