--- packages/dds/tree/src/class-tree/internal.ts
@@ -1,20 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-// Used by public types, but not part of the desired API surface
-
-export {
-	ObjectFromSchemaRecord,
-	InsertableObjectFromSchemaRecord,
-	TreeNodeFromImplicitAllowedTypes,
-	ImplicitAllowedTypes,
-	FieldKind,
-	InsertableTreeFieldFromImplicitField,
-	AllowedTypes,
-	FieldSchema,
-	ApplyKind,
-	InsertableTypedNode,
-	NodeBuilderData,
-} from "./schemaTypes";
