+++ packages/dds/tree/src/feature-libraries/index.ts
@@ -135,7 +135,6 @@ export {
 	Any,
 	SchemaLibraryData,
 	LazyTreeNodeSchema,
-	InternalTypedSchemaTypes,
 	ViewSchema,
 	SchemaLintConfiguration,
 	FieldNodeSchema,
