+++ packages/dds/tree/src/feature-libraries/typed-schema/index.ts
@@ -55,7 +55,3 @@ export {
 } from "./flexList";
 
 export { ArrayToUnion } from "./typeUtils";
-
-// Below here are things that are used by the above, but not part of the desired API surface.
-import * as InternalTypedSchemaTypes from "./internal";
-export { InternalTypedSchemaTypes };
