--- packages/dds/tree/src/feature-libraries/typed-schema/internal.ts
@@ -1,10 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-// Used by public types, but not part of the desired API surface
-
-export { ArrayToUnion } from "./typeUtils";
-
-export { FlexList, LazyItem, ExtractItemType, FlexListToUnion } from "./flexList";
