--- packages/dds/tree/src/internal.ts
@@ -1,17 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-// Used by public types, but not part of the desired API surface.
-export { MakeNominal, RestrictiveReadonlyRecord } from "./util";
-
-export {
-	NodeKind,
-	TreeNodeSchemaClass,
-	TreeNodeSchemaNonClass,
-	TreeNodeSchemaCore,
-	InternalClassTreeTypes,
-} from "./class-tree";
-
-export { TreeArrayNodeBase } from "./simple-tree";
