+++ packages/dds/tree/src/simple-tree/types.ts
@@ -12,12 +12,12 @@ import {
 	Any,
 	FieldNodeSchema,
 	TreeFieldSchema,
-	InternalTypedSchemaTypes,
 	LeafNodeSchema,
 	MapNodeSchema,
 	ObjectNodeSchema,
 	TreeNodeSchema,
 	AssignableFieldKinds,
+	LazyItem,
 } from "../feature-libraries";
 import { IterableTreeListContent, TreeListNodeOld } from "./treeListNode";
 
@@ -338,9 +338,7 @@ export type TreeNodeUnion<TTypes extends AllowedTypes> = TTypes extends readonly
 			// TODO: Is the the best way to write this type function? Can it be simplified?
 			// This first maps the tuple of AllowedTypes to a tuple of node API types.
 			// Then, it uses [number] to index arbitrarily into that tuple, effectively converting the type tuple into a type union.
-			[Index in keyof TTypes]: TTypes[Index] extends InternalTypedSchemaTypes.LazyItem<
-				infer InnerType
-			>
+			[Index in keyof TTypes]: TTypes[Index] extends LazyItem<infer InnerType>
 				? InnerType extends TreeNodeSchema
 					? TypedNode<InnerType>
 					: never
